+++ src/features/auth/ui/SignUpButton.tsx
@@ -1,0 +1,7 @@
+import { Button } from "shared/ui/button"
+
+export const SignUpButton = () => {
+    return (
+        <Button>Sign up</Button>
+    )
+}